refs #148 Параллельные и последовательные варианты сортировки подсчётом пусть лежат в своих файлах
